split tests to separate file
